+++ include/registration_est_bingham_kf_rgbd.h
@@ -10,11 +10,11 @@ struct RegistrationResult{
 };
 
 extern "C" struct RegistrationResult *registration_est_bingham_kf_rgbd(PointCloud *ptcldMoving, PointCloud *ptcldFixed,
-                                                                       int inlierRatio, int maxIterations, int windowSize,
+                                                                       double inlierRatio, int maxIterations, int windowSize,
                                                                        double toleranceT, double toleranceR);
 
 extern "C" struct RegistrationResult *registration_est_bingham_normal(PointCloud *ptcldMoving, PointCloud *ptcldFixed,
                                                            PointCloud *normalMoving, PointCloud *normalFixed,
-                                                           int inlierRatio, int maxIterations, int windowSize,
+                                                           double inlierRatio, int maxIterations, int windowSize,
                                                            double toleranceT, double toleranceR);
 #endif
